requirements.txt
streamlit
tensorflow
numpy
Pillow
matplotlib
seaborn
scikit-learn
pandas
requests